Rename component file names to proper names for clarity
