+++ src/components/CharacterCard/CharacterCard.tsx
